+++ src/main/java/net/brdle/collectorsreap/compat/JEIPlugin.java
@@ -20,8 +20,8 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.fluids.FluidStack;
 import java.util.ArrayList;
 import java.util.List;
-import javax.annotation.ParametersAreNonnullByDefault;
 import net.minecraftforge.registries.RegistryObject;
+import javax.annotation.ParametersAreNonnullByDefault;
 import org.jetbrains.annotations.NotNull;
 
 @JeiPlugin